requirements.txt
# Formula 1 Schedule App Dependencies
fastmcp>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
typing-extensions>=4.7.0